Group fixtures in conftest
